# Revert "Optimise merge registers for migrations"