--- a/cmd/util/ledger/migrations/account_storage_migration.go
+++ b/cmd/util/ledger/migrations/account_storage_migration.go
@@ -58,9 +58,11 @@ func NewAccountStorageMigration(
 			flow.Address(address): {},
 		}
 
-		newPayloads, err := migrationRuntime.Snapshot.ApplyChangesAndGetNewPayloads(
+		newPayloads, err := MergeRegisterChanges(
+			migrationRuntime.Snapshot.Payloads,
 			result.WriteSet,
 			expectedAddresses,
+			nil,
 			log,
 		)
 		if err != nil {