--- a/cmd/util/ledger/migrations/cadence_values_migration.go
+++ b/cmd/util/ledger/migrations/cadence_values_migration.go
@@ -145,8 +145,10 @@ func (m *CadenceBaseMigrator) MigrateAccount(
 		flow.Address(address): {},
 	}
 
-	newPayloads, err := migrationRuntime.Snapshot.ApplyChangesAndGetNewPayloads(
+	newPayloads, err := MergeRegisterChanges(
+		migrationRuntime.Snapshot.Payloads,
 		result.WriteSet,
+		expectedAddresses,
 		expectedAddresses,
 		m.log,
 	)