--- a/cmd/util/ledger/migrations/cadence_values_migration_test.go
+++ b/cmd/util/ledger/migrations/cadence_values_migration_test.go
@@ -796,9 +796,11 @@ func TestProgramParsingError(t *testing.T) {
 		flow.Address(testAddress): {},
 	}
 
-	payloads, err = runtime.Snapshot.ApplyChangesAndGetNewPayloads(
+	payloads, err = MergeRegisterChanges(
+		runtime.Snapshot.Payloads,
 		result.WriteSet,
 		expectedAddresses,
+		nil,
 		logger,
 	)
 	require.NoError(t, err)
@@ -932,9 +934,11 @@ func TestCoreContractUsage(t *testing.T) {
 			flow.Address(testAddress): {},
 		}
 
-		payloads, err = runtime.Snapshot.ApplyChangesAndGetNewPayloads(
+		payloads, err = MergeRegisterChanges(
+			runtime.Snapshot.Payloads,
 			result.WriteSet,
 			expectedAddresses,
+			nil,
 			logger,
 		)
 		require.NoError(t, err)