--- a/cmd/util/ledger/util/util.go
+++ b/cmd/util/ledger/util/util.go
@@ -12,6 +12,7 @@ import (
 	"github.com/onflow/cadence/runtime/common"
 
 	"github.com/onflow/flow-go/fvm/environment"
+	"github.com/onflow/flow-go/fvm/storage/snapshot"
 	"github.com/onflow/flow-go/ledger"
 	"github.com/onflow/flow-go/ledger/common/convert"
 	"github.com/onflow/flow-go/model/flow"
@@ -68,6 +69,38 @@ func (a *AccountsAtreeLedger) AllocateStorageIndex(owner []byte) (atree.StorageI
 	return v, nil
 }
 
+type PayloadSnapshot struct {
+	Payloads map[flow.RegisterID]*ledger.Payload
+}
+
+var _ snapshot.StorageSnapshot = (*PayloadSnapshot)(nil)
+
+func NewPayloadSnapshot(payloads []*ledger.Payload) (*PayloadSnapshot, error) {
+	l := &PayloadSnapshot{
+		Payloads: make(map[flow.RegisterID]*ledger.Payload, len(payloads)),
+	}
+	for _, payload := range payloads {
+		key, err := payload.Key()
+		if err != nil {
+			return nil, err
+		}
+		id, err := convert.LedgerKeyToRegisterID(key)
+		if err != nil {
+			return nil, err
+		}
+		l.Payloads[id] = payload
+	}
+	return l, nil
+}
+
+func (p PayloadSnapshot) Get(id flow.RegisterID) (flow.RegisterValue, error) {
+	value, exists := p.Payloads[id]
+	if !exists {
+		return nil, nil
+	}
+	return value.Value(), nil
+}
+
 // NopMemoryGauge is a no-op implementation of the MemoryGauge interface
 type NopMemoryGauge struct{}
 
@@ -78,7 +111,7 @@ func (n NopMemoryGauge) MeterMemory(common.MemoryUsage) error {
 var _ common.MemoryGauge = (*NopMemoryGauge)(nil)
 
 type PayloadsReadonlyLedger struct {
-	Snapshot MigrationStorageSnapshot
+	Snapshot *PayloadSnapshot
 
 	AllocateStorageIndexFunc func(owner []byte) (atree.StorageIndex, error)
 	SetValueFunc             func(owner, key, value []byte) (err error)
@@ -100,9 +133,9 @@ func (p *PayloadsReadonlyLedger) SetValue(owner, key, value []byte) (err error) 
 	panic("SetValue not expected to be called")
 }
 
-func (p *PayloadsReadonlyLedger) ValueExists(owner, key []byte) (bool, error) {
-	exists := p.Snapshot.Exists(flow.NewRegisterID(flow.BytesToAddress(owner), string(key)))
-	return exists, nil
+func (p *PayloadsReadonlyLedger) ValueExists(owner, key []byte) (exists bool, err error) {
+	_, ok := p.Snapshot.Payloads[flow.NewRegisterID(flow.BytesToAddress(owner), string(key))]
+	return ok, nil
 }
 
 func (p *PayloadsReadonlyLedger) AllocateStorageIndex(owner []byte) (atree.StorageIndex, error) {
@@ -113,7 +146,7 @@ func (p *PayloadsReadonlyLedger) AllocateStorageIndex(owner []byte) (atree.Stora
 	panic("AllocateStorageIndex not expected to be called")
 }
 
-func NewPayloadsReadonlyLedger(snapshot MigrationStorageSnapshot) *PayloadsReadonlyLedger {
+func NewPayloadsReadonlyLedger(snapshot *PayloadSnapshot) *PayloadsReadonlyLedger {
 	return &PayloadsReadonlyLedger{Snapshot: snapshot}
 }