… cmd/util/ledger/util/payload_snapshot.go
… cmd/util/ledger/util/payload_snapshot_test.go